+++ app/src/main/java/com/ft/DemoApplication.java
@@ -1,23 +1,15 @@
 package com.ft;
 
-import android.app.Activity;
 import android.content.Context;
 
-import com.ft.sdk.ActionEventWrapper;
 import com.ft.sdk.DeviceMetricsMonitorType;
 import com.ft.sdk.EnvType;
 import com.ft.sdk.ErrorMonitorType;
-import com.ft.sdk.FTActionTrackingHandler;
 import com.ft.sdk.FTLoggerConfig;
 import com.ft.sdk.FTRUMConfig;
 import com.ft.sdk.FTSDKConfig;
 import com.ft.sdk.FTSdk;
 import com.ft.sdk.FTTraceConfig;
-import com.ft.sdk.FTViewActivityTrackingHandler;
-import com.ft.sdk.FTViewFragmentTrackingHandler;
-import com.ft.sdk.FragmentWrapper;
-import com.ft.sdk.HandlerAction;
-import com.ft.sdk.HandlerView;
 import com.ft.sdk.LogCacheDiscard;
 import com.ft.sdk.RUMCacheDiscard;
 import com.ft.sdk.TraceType;
@@ -79,7 +71,7 @@ public class DemoApplication extends BaseApplication {
 
 
     static void initFTSDK(Context context) {
-        FTSDKConfig ftSDKConfig = FTSDKConfig.builder(BuildConfig.DATAKIT_URL)
+        FTSDKConfig ftSDKConfig = FTSDKConfig.builder(BuildConfig.DATAWAY_URL, BuildConfig.CLIENT_TOKEN)
                 .setDebug(true)//Set whether it's debug
                 .setAutoSync(true)
                 .setCustomSyncPageSize(10)