+++ ft-session-replay/CHANGELOG.md
@@ -1,3 +1,7 @@
+# replay 0.1.3-alpha01
+1. Webview Session Replay support 0.1.3-alpha01
+
+---
 # replay 0.1.2-alpha02
 1. Adapt to react native replay feature